+++ extract_msg/structures/recurrence_pattern.py
@@ -60,7 +60,7 @@ class RecurrencePattern:
         return self.__calendarType
 
     @property
-    def deletedInstanceDates(self) -> Tuple[int]:
+    def deletedInstanceDates(self) -> Tuple[int, ...]:
         """
         A tuple of the dates (stored as number of minutes between midnight,
         January 1, 1601, and midnight on the specified day in the timezone
@@ -104,7 +104,7 @@ class RecurrencePattern:
         return self.__firstDOW
 
     @property
-    def modifiedInstanceDates(self) -> Tuple[int]:
+    def modifiedInstanceDates(self) -> Tuple[int, ...]:
         """
         A tuple of the dates (stored as number of minutes between midnight,
         January 1, 1601, and midnight on the specified day in the timezone